Use 'double' for cosineSimilarity in NNQueryResult and JNI binding

+++ include/NNQueryResult.h
@@ -6,7 +6,7 @@
 
 struct NNQueryResult {
     std::string personName;
-    float cosineSimilarity = 0.0f;
+    double cosineSimilarity = 0.0;
     FaceBoundingBox faceBoundingBox;
 };
